back/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==3.4.7
Django==4.2.16
django-allauth==0.54.0
django-cors-headers==3.14.0
dj-rest-auth==2.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
oauthlib==3.2.2
outcome==1.3.0.post0
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
